+++ src/ServiceProvider.php
@@ -1,4 +1,4 @@
-<?php namespace DaveJamesMiller\Breadcrumbs;
+<?php namespace Laracraft\Breadcrumbs;
 
 use Illuminate\Container\Container;
 use Illuminate\Support\ServiceProvider as BaseServiceProvider;
@@ -32,19 +32,17 @@ class ServiceProvider extends BaseServiceProvider {
 	 */
 	public function register()
 	{
-	    $this->app->bind('DaveJamesMiller\Breadcrumbs\Generator', function ($app) {
+	    $this->app->bind('Laracraft\Breadcrumbs\Generator', function ($app) {
 	        return new Generator(new Container(), $app['config']['breadcrumbs.namespace']);
         });
 
 		$this->app['breadcrumbs'] = $this->app->share(function($app)
 		{
-			$breadcrumbs = $this->app->make('DaveJamesMiller\Breadcrumbs\Manager');
+			$breadcrumbs = $this->app->make('Laracraft\Breadcrumbs\Manager');
 
 			$viewPath = __DIR__ . '/../views/';
 
 			$this->loadViewsFrom($viewPath, 'breadcrumbs');
-			$this->loadViewsFrom($viewPath, 'laravel-breadcrumbs'); // Backwards-compatibility with 2.x
-
 			$breadcrumbs->setView($app['config']['breadcrumbs.view']);
 
 			return $breadcrumbs;