+++ tests/TestCase.php
@@ -5,14 +5,14 @@ abstract class TestCase extends Orchestra\Testbench\TestCase {
 	protected function getPackageProviders($app)
 	{
 		return [
-			DaveJamesMiller\Breadcrumbs\ServiceProvider::class,
+			Laracraft\Breadcrumbs\ServiceProvider::class,
 		];
 	}
 
 	protected function getPackageAliases($app)
 	{
 		return [
-			'Breadcrumbs' => DaveJamesMiller\Breadcrumbs\Facade::class
+			'Breadcrumbs' => Laracraft\Breadcrumbs\Facade::class
 		];
 	}