+++ tests/integration/DependantServiceProviderTest.php
@@ -5,7 +5,7 @@ class DependantServiceProviderTest extends TestCase {
 	protected function getPackageProviders($app)
 	{
 		return [
-			DaveJamesMiller\Breadcrumbs\ServiceProvider::class,
+			Laracraft\Breadcrumbs\ServiceProvider::class,
 			DependantServiceProvider::class,
 		];
 	}