+++ tests/unit/CurrentRouteTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use DaveJamesMiller\Breadcrumbs\CurrentRoute;
+use Laracraft\Breadcrumbs\CurrentRoute;
 use Mockery as m;
 
 class CurrentRouteTest extends TestCase {
@@ -9,7 +9,7 @@ class CurrentRouteTest extends TestCase {
 	{
 		parent::setUp();
 
-		$this->currentRoute = app('DaveJamesMiller\Breadcrumbs\CurrentRoute');
+		$this->currentRoute = app('Laracraft\Breadcrumbs\CurrentRoute');
 	}
 
 	public function testNamedRoute()
@@ -39,7 +39,7 @@ class CurrentRouteTest extends TestCase {
 	}
 
     /**
-     * @expectedException DaveJamesMiller\Breadcrumbs\Exception
+     * @expectedException Laracraft\Breadcrumbs\Exception
      * @expectedExceptionMessage The current route (GET /sample/unnamed) is not named - please check routes.php for an "as" parameter
      */
 	public function testUnnamedRoute()