+++ tests/unit/GeneratorTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use DaveJamesMiller\Breadcrumbs\Generator;
+use Laracraft\Breadcrumbs\Generator;
 use Mockery as m;
 
 class GeneratorTest extends TestCase {